Made changes to leaderboard
